Implement backwards-compatibility
